TowStyTemplates.jl

Templates for TowSty, the webapp and static site generator for Stylo.

TowStyTemplates.jl is mainly based on FranklinTemplates.jl.

Getting started

pkg> add TowStyTemplates

julia> using TowStyTemplates
julia> newproject("myProject", template="jj")

List of templates

licence	name	navbar

GPL 3.0	jj	true

GPL 3.0	tinlizzie	true

API

TowStyTemplates.generatehash — Method

generatehash(dir::String)

This function generates a hash and writes it on a file named .hash for workspace update.

TowStyTemplates.mergefolders — Method

mergefolders(src, dst)

Internal function to looks at what's inside src/ and put it in dst/. If there are paths that match, the files are merged. It is assumed that files will not clash, if they clash files in dst are replaced.

TowStyTemplates.newproject — Function

newproject(dir; template="jj", changedir=true)

Generate a new folder (an error is thrown if it already exists) that contains the skeleton of the blog that can be processed by TowSty. The user can specify a template out of the list of available templates. If topdir is specified as "." then the current directory is used.

template="jj": name of the template to use,

changedir=true: whether to change the current directory to the newly

created folder or not.

verbose=true: whether to display information or not.

Example

newproject("MyNewWebsite", template="jj")

TowStyTemplates.templates — Method

templates()

This function lists available templates.